Show an alert message before start the firmware upgrade process
Some boards exposes the strapping pins, and if one of them don't have the expected logical value when the board is reset the firmware upgrade process can be affected.

The following alert message has been added before start the upgrade process:

"Please, remove any hardware attached to your board before upgrading the firmware.

This is required because some boards exposes the strapping pins, which are special pins that can be interfere in the upgrade process if they not have the expected logical value when the board is reset."